--- a/zipkin-web/build.gradle
+++ b/zipkin-web/build.gradle
@@ -35,9 +35,6 @@ node {
     // Version of node to use.
     version = '5.5.0'
 
-    // Version of npm to use.
-    npmVersion = '3.6.0'
-
     // Base URL for fetching node distributions (change if you have a mirror).
     distBaseUrl = 'https://nodejs.org/dist'