@eirslett 2016-02-25
Can gradle check file diffs as well? Given that package.json (and shrinkwrap.json) isn't changed, we shouldn't need to invoke npm at all. The same goes for the webpack build.