@abesto 2016-02-26
Yes, but consider: if package.json didn't change, but i manually removed `node_modules`, then you'd expect gradle to run `npm install`. Same if I just removed one package manually. AFAICT the plugin handles this by considering the whole node_modules dir for diffs, as evidenced by https://github.com/srs/gradle-node-plugin/issues/94; I don't have a good idea for improving things.